+++ src/main/java/com/demo/udema/service/LessonService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface LessonService {
     Lessons findByTitle(String title);
     Lessons findById(int id);
-    Double findLessonsSumByTitle(String title);
+    Integer findLessonsSumByTitle(String title);
     Integer countLessonsByTitle(String title);
 }
